Finished Main and made some changes for manual vs CV play.
I "finished" the main function but it needs to be debugged. I also reset
the functions from Matlab indexing to C++ indexing so player1 (us) is 0
and p2 is 1 etc. I added some things for passing in stuff from the CV
system. Finally, I made whos_turn a global variable to simplify things.